WIP - ps3 removed, Joystick added
I now have a copy of my joystick code in this project and removed the
ps3 version.  I believe all of the code compiles and I verified that I
could start the ps3 demo without running sudo.  Still need to testing to
see how everything works and how best to update startup script for this.

also updated RME code, but so far the places that call off to joystick
are commented out.  Need to find out why.