WIP: Support Serial port access to Arbotix Pro
Add support for allowing me to experiment with connecting up my modified
firmware for the Arbotix Pro, which allows me to use other TTL serial
ports in addition to USB to talk to Arbotix Pro.

Looks first for /dev/ttyDXL as an override for the passed in string,
which is probably always /dev/ttyUSB0.  If found uses it.

Also trying to handle issues where some of the different terminal
drivers like FTDI and ACM and... Don't all handle setting the baud rate
the same way, nor do they handle tcdrain the same.